v2/batch: remove concurrent_worker
It's been completely replaced by concurrent_chan_worker. Removing this
will avoid useless changes to concurrent_worker anytime we want to make
changes to the batch interface.
