+++ v2/internal/pkg/batch/new.go
@@ -1,7 +1,6 @@
 package batch
 
 import (
-	"github.com/openshift/oc-mirror/v2/internal/pkg/api/v2alpha1"
 	clog "github.com/openshift/oc-mirror/v2/internal/pkg/log"
 	"github.com/openshift/oc-mirror/v2/internal/pkg/mirror"
 )
@@ -12,16 +11,5 @@ func New(workerType string,
 	mirror mirror.MirrorInterface,
 	batchSize uint,
 ) BatchInterface {
-	copiedImages := v2alpha1.CollectorSchema{
-		AllImages: []v2alpha1.CopyImageSchema{},
-	}
-
-	switch workerType {
-	case ConcurrentWorker:
-		return &ConcurrentBatch{Log: log, LogsDir: logsDir, Mirror: mirror, CopiedImages: copiedImages, BatchSize: batchSize}
-	case ChannelConcurrentWorker:
-		return &ChannelConcurrentBatch{Log: log, LogsDir: logsDir, Mirror: mirror, MaxGoroutines: int(batchSize)}
-	default:
-		return &ChannelConcurrentBatch{Log: log, LogsDir: logsDir, Mirror: mirror, MaxGoroutines: int(batchSize)}
-	}
+	return &ChannelConcurrentBatch{Log: log, LogsDir: logsDir, Mirror: mirror, MaxGoroutines: int(batchSize)}
 }
